third_party/WebKit/LayoutTests/bluetooth/characteristic/notifications/event-after-starting.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('heart_rate'))
    .then(service => service.getCharacteristic('heart_rate_measurement'))
    .then(characteristic => {
      return assert_event_fires_after_promise(
        characteristic, 'startNotifications', 'characteristicvaluechanged',
        1 /* attach 1 listener */, false /* ignore_event_promise_order */)
        .then(() => characteristic.stopNotifications())
        .then(() => assert_no_events(characteristic,
                                     'characteristicvaluechanged'));
    });
}, 'Starting notifications should fire an event and we should get no events ' +
   'after notifications are stopped.');
</script>
